requirement.txt
fastapi 
uvicorn 
pdf2image 
opencv-python-headless 
pillow 
numpy